Address nikic's comments
