+++ llvm/lib/IR/Globals.cpp
@@ -139,8 +139,6 @@ void GlobalObject::copyAttributesFrom(const GlobalObject *Src) {
   GlobalValue::copyAttributesFrom(Src);
   setAlignment(Src->getAlign());
   setSection(Src->getSection());
-  if (auto CM = Src->getCodeModel())
-    setCodeModel(*CM);
 }
 
 std::string GlobalValue::getGlobalIdentifier(StringRef Name,
@@ -263,15 +261,6 @@ void GlobalObject::setSection(StringRef S) {
   // Update the HasSectionHashEntryBit. Setting the section to the empty string
   // means this global no longer has a section.
   setGlobalObjectFlag(HasSectionHashEntryBit, !S.empty());
-}
-
-void GlobalObject::setCodeModel(CodeModel::Model CM) {
-  unsigned CodeModelData = static_cast<unsigned>(CM) + 1;
-  unsigned OldData = getGlobalValueSubClassData();
-  unsigned NewData = (OldData & ~(CodeModelMask << CodeModelShift)) |
-                     (CodeModelData << CodeModelShift);
-  setGlobalValueSubClassData(NewData);
-  assert(getCodeModel() == CM && "Code model representation error!");
 }
 
 bool GlobalValue::isNobuiltinFnDef() const {
@@ -493,11 +482,22 @@ void GlobalVariable::copyAttributesFrom(const GlobalVariable *Src) {
   GlobalObject::copyAttributesFrom(Src);
   setExternallyInitialized(Src->isExternallyInitialized());
   setAttributes(Src->getAttributes());
+  if (auto CM = Src->getCodeModel())
+    setCodeModel(*CM);
 }
 
 void GlobalVariable::dropAllReferences() {
   User::dropAllReferences();
   clearMetadata();
+}
+
+void GlobalVariable::setCodeModel(CodeModel::Model CM) {
+  unsigned CodeModelData = static_cast<unsigned>(CM) + 1;
+  unsigned OldData = getGlobalValueSubClassData();
+  unsigned NewData = (OldData & ~(CodeModelMask << CodeModelShift)) |
+                     (CodeModelData << CodeModelShift);
+  setGlobalValueSubClassData(NewData);
+  assert(getCodeModel() == CM && "Code model representation error!");
 }
 
 //===----------------------------------------------------------------------===//
